Create a circular area-sink analytic element with specified recharge

areasink() creates a circular area-sink analytic element with constant, uniform specified recharge.

Usage

areasink(xc, yc, N, R, location = c("top", "base"), ...)

Arguments

xc

numeric, x location of the center of the area-sink

yc

numeric, y location of the center of the area-sink

N

numeric, uniform constant leakage value (positive is into aquifer) in length per time.

R

numeric, radius of the circular area-sink

location

character, either top (default) or base specifying the vertical position of the area-sink.

...

ignored

Value

Circular area-sink analytic element which is an object of class areasink and inherits from element.

Details

Area-sinks can be used to simulate areal recharge or seepage at the aquifer top, or leakage into or out of the aquifer at its base. The location argument is used when calculating the vertical flow component.

Examples

areasink(xc = -500, yc = 0, N = 0.001, R = 500)
#> $parameter
#> [1] 0.001
#> 
#> $nunknowns
#> [1] 0
#> 
#> $xc
#> [1] -500
#> 
#> $yc
#> [1] 0
#> 
#> $R
#> [1] 500
#> 
#> $location
#> [1] "top"
#> 
#> attr(,"class")
#> [1] "areasink" "element" 

# flux assuming a constant head difference over a confining unit
dh <- 3
res <- 10 / 0.0001
areasink(xc = -500, yc = 0, N = -dh/res, R = 500, location = 'base')
#> $parameter
#> [1] -3e-05
#> 
#> $nunknowns
#> [1] 0
#> 
#> $xc
#> [1] -500
#> 
#> $yc
#> [1] 0
#> 
#> $R
#> [1] 500
#> 
#> $location
#> [1] "base"
#> 
#> attr(,"class")
#> [1] "areasink" "element"