Create a head-specified line-sink analytic element

headlinesink() creates a line-sink analytic element with constant specified head. The discharge into the line-sink per unit length is computed by solving the corresponding aem model.

Usage

headlinesink(x0, y0, x1, y1, hc, resistance = 0, width = 0, ...)

Arguments

x0

numeric, starting x location of line-sink.

y0

numeric, starting y location of line-sink.

x1

numeric, ending x location of line-sink.

y1

numeric, ending y location of line-sink.

hc

numeric, specified hydraulic head of the line-sink.

resistance

numeric, hydraulic resistance of the line-sink at its connection with the aquifer. Defaults to 0 (no resistance).

width

numeric, width of the line-sink. Used with resistance to calculate the line-sink strength, and by tracelines() to determine if a particle has reached the line. Defaults to zero (infinitesimally narrow line).

...

ignored

Value

Head-specified line-sink analytic element which is an object of class headlinesink and inherits from linesink.

Details

The strength of the line-sink (discharge per unit length of line-sink) is computed by solving the aem model given the specified head hc for the line-sink. This head is located at the so-called collocation point, which is placed at the center of the line-sink.

The resistance can be increased for a line-sink in poor connection with the aquifer. The effect of a larger or smaller wetted perimeter can be mimicked by adjusting the resistance and/or width accordingly. If width = 0 (the default) it is removed from the conductance calculation. If the aquifer is unconfined (i.e. has a variable saturated thickness), the system of equations becomes non-linear with respect to the hydraulic head and iteration is required to solve the model.

See also

Examples

hls <- headlinesink(-75, 50, 100, 50, hc = 10)
hls <- headlinesink(-75, 50, 100, 50, hc = 10, resistance = 10, width = 4)