Create a uniform-flow analytic element with specified recharge

uniformflow() creates a analytic element of constant uniform background flow.

Usage

uniformflow(TR, gradient, angle, ...)

Arguments

TR

numeric, constant transmissivity value used to define the discharge.

gradient

numeric, hydraulic gradient. Positive in the direction of flow.

angle

numeric, angle of the primary direction of background flow, in degrees counterclockwise from the x-axis.

...

ignored

Value

Analytic element of constant uniform flow which is an object of class uniformflow and inherits from element.

Details

TR and gradient are multiplied to obtain the discharge which remains constant throughout the system, independent of the saturated thickness of the aquifer.

Groundwater flow is in the direction of the negative hydraulic gradient. Note that gradient is specified here as positive in the direction of flow for convenience.

Examples

uniformflow(TR = 100, gradient = 0.002, angle = -45) # South-eastern direction
#> $parameter
#> [1] 0.2
#> 
#> $nunknowns
#> [1] 0
#> 
#> $udir
#> [1] 0.7071068+0.7071068i
#> 
#> attr(,"class")
#> [1] "uniformflow" "element"