Create an analytic element of a constant-discharge well

well() creates an analytic element of a well with constant discharge.

Usage

well(xw, yw, Q, rw = 0.3, ...)

Arguments

xw

numeric, x location of the well.

yw

numeric, y location of the well.

Q

numeric, volumetric discharge of the well (positive is out of aquifer).

rw

numeric, radius of well. Defaults to 0.3 length units.

...

ignored

Value

Analytic element of a well with constant discharge which is an object of class well and inherits from element.

Details

The inner annulus of a well element constitutes a singularity in the equations as the hydraulic head is undefined at a distance smaller than rw from the well center. If a state- or flow-variable is calculated within this annulus, its location is reset to its nearest location on the well screen.

The well is assumed to fully penetrate the saturated aquifer.

See also

Examples

w <- well(xw = 50, yw = 0, Q = 200, rw = 0.3)